requirements.txt
sounddevice
numpy
scipy
groq
pandas
openpyxl
python-dotenv
pyyaml